Add a channel stream and use it with the RMS and STD
