+++ src/STDModule.hpp
@@ -9,33 +9,15 @@
 #define DQM_SRC_STDMODULE_HPP_
 
 // DQM
-#include "dqm/AnalysisModule.hpp"
-#include "ChannelMap.hpp"
-#include "Constants.hpp"
-#include "Decoder.hpp"
-#include "Exporter.hpp"
-#include "dqm/Issues.hpp"
 #include "dqm/algs/STD.hpp"
-#include "dqm/FormatUtils.hpp"
-#include "dqm/Pipeline.hpp"
-#include "dqm/DQMFormats.hpp"
-#include "dqm/DQMLogging.hpp"
+#include "ChannelStream.hpp"
 
-#include "daqdataformats/TriggerRecord.hpp"
-#include "detdataformats/tde/TDE16Frame.hpp"
-
-#include <cstdlib>
-#include <map>
-#include <memory>
 #include <string>
 #include <vector>
-#include <chrono>
 
 namespace dunedaq::dqm {
 
-using logging::TLVL_WORK_STEPS;
-
-class STDModule : public AnalysisModule
+class STDModule : public ChannelStream<STD>
 {
 
 public:
@@ -43,32 +25,6 @@ public:
             int nchannels,
             std::vector<int>& link_idx);
 
-  void run(std::shared_ptr<daqdataformats::TriggerRecord> record,
-      DQMArgs& args, DQMInfo& info) override;
-
-  template <class T>
-  void run_(std::shared_ptr<daqdataformats::TriggerRecord> record,
-       DQMArgs& args, DQMInfo& info);
-
-  // std::unique_ptr<daqdataformats::TriggerRecord>
-  // run_tdeframe(std::unique_ptr<daqdataformats::TriggerRecord> record,
-  //               std::shared_ptr<ChannelMap>& map,
-  //               const std::string& kafka_address = "");
-  void transmit(const std::string& kafka_address,
-                std::shared_ptr<ChannelMap>& map,
-                const std::string& topicname,
-                int run_num);
-
-  void clean();
-  void fill(int ch, double value);
-  void fill(int ch, int link, double value);
-  int get_local_index(int ch, int link);
-
-private:
-  std::string m_name;
-  std::vector<STD> histvec;
-  int m_size;
-  std::map<int, int> m_index;
 };
 
 
@@ -76,170 +32,10 @@ STDModule::STDModule(std::string name,
                              int nchannels,
                              std::vector<int>& link_idx
                      )
-  : m_name(name)
-  , m_size(nchannels)
+  : ChannelStream(name, nchannels, link_idx,
+                  [this] (std::vector<STD>& vec, int ch, int link) {
+                    return vec[get_local_index(ch, link)].std();})
 {
-  for (int i = 0; i < m_size; ++i) {
-    histvec.emplace_back(STD());
-  }
-  int channels = 0;
-  for (size_t i = 0; i < link_idx.size(); ++i) {
-    m_index[link_idx[i]] = channels;
-    channels += CHANNELS_PER_LINK;
-  }
-}
-
-// std::unique_ptr<daqdataformats::TriggerRecord>
-// STDModule::run_tdeframe(std::unique_ptr<daqdataformats::TriggerRecord> record,
-//                              std::shared_ptr<ChannelMap>& map,
-//                              const std::string& kafka_address)
-// {
-//   TLOG() << "Running run_tdeframe";
-//   auto wibframes = decode<detdataformats::tde::TDE16Frame>(*record);
-//   return std::move(record);
-// }
-
-template <class T>
-void
-STDModule::run_(std::shared_ptr<daqdataformats::TriggerRecord> record,
-                DQMArgs& args, DQMInfo& info)
-{
-  auto map = args.map;
-
-  auto frames = decode<T>(record, args.max_frames);
-  auto pipe = Pipeline<T>({"remove_empty", "check_empty", "check_timestamps_aligned"});
-  bool valid_data = pipe(frames);
-  if (!valid_data) {
-    return;
-  }
-
-  // Get all the keys
-  std::vector<int> keys;
-  for (auto& [key, value] : frames) {
-    keys.push_back(key);
-  }
-
-  for (const auto& [key, vec] : frames) {
-    for (const auto& fr : vec) {
-      for (int ich = 0; ich < CHANNELS_PER_LINK; ++ich) {
-        fill(ich, key, get_adc<T>(fr, ich));
-      }
-    }
-  }
-
-  transmit(args.kafka_address,
-           map,
-           args.kafka_topic,
-           record->get_header_ref().get_run_number());
-  clean();
-
-
-}
-
-
-
-void
-STDModule::run(std::shared_ptr<daqdataformats::TriggerRecord> record,
-               DQMArgs& args, DQMInfo& info)
-{
-  TLOG(TLVL_WORK_STEPS) << "Running STD with frontend_type = " << args.frontend_type;
-  auto start = std::chrono::steady_clock::now();
-  auto frontend_type = args.frontend_type;
-  if (frontend_type == "wib") {
-    set_is_running(true);
-    run_<detdataformats::wib::WIBFrame>(std::move(record), args, info);
-    set_is_running(false);
-  }
-  else if (frontend_type == "wib2") {
-    set_is_running(true);
-    run_<detdataformats::wib2::WIB2Frame>(std::move(record), args, info);
-    set_is_running(false);
-  }
-  auto stop = std::chrono::steady_clock::now();
-
-
-  // else if (frontend_type == "tde") {
-  //   set_is_running(true);
-  //   auto ret = run_<detdataformats::wib::WIBFrame>(std::move(record), map, kafka_address);
-  //   set_is_running(false);
-  //   return ret;
-  // }
-  info.std_time_taken.store(std::chrono::duration_cast<std::chrono::milliseconds>(stop-start).count());
-  info.std_times_run++;
-}
-
-void
-STDModule::transmit(const std::string& kafka_address,
-                    std::shared_ptr<ChannelMap>& cmap,
-                    const std::string& topicname,
-                    int run_num)
-{
-  // Placeholders
-  std::string dataname = m_name;
-  std::string partition = getenv("DUNEDAQ_PARTITION");
-  std::string app_name = getenv("DUNEDAQ_APPLICATION_NAME");
-  std::string datasource = partition + "_" + app_name;
-
-  // One message is sent for every plane
-  auto channel_order = cmap->get_map();
-  for (auto& [plane, map] : channel_order) {
-    std::stringstream output;
-    output << "{";
-    output << "\"source\": \"" << datasource << "\",";
-    output << "\"run_number\": \"" << run_num << "\",";
-    output << "\"partition\": \"" << partition << "\",";
-    output << "\"app_name\": \"" << app_name << "\",";
-    output << "\"plane\": \"" << plane << "\",";
-    output << "\"algorithm\": \"" << "std" << "\"";
-    output << "}\n\n\n";
-    std::vector<int> channels;
-    for (auto& [offch, pair] : map) {
-      channels.push_back(offch);
-    }
-    auto bytes = serialization::serialize(channels, serialization::kMsgPack);
-    for (auto& b : bytes) {
-      output << b;
-    }
-    output << "\n\n\n";
-    std::vector<float> values;
-    for (auto& [offch, pair] : map) {
-      int link = pair.first;
-      int ch = pair.second;
-      values.push_back(histvec[get_local_index(ch, link)].std());
-    }
-    bytes = serialization::serialize(values, serialization::kMsgPack);
-    for (auto& b : bytes) {
-      output << b;
-    }
-    TLOG(TLVL_WORK_STEPS) << "Sending message to kafka with size " << output.str().size();
-    KafkaExport(kafka_address, output.str(), topicname);
-  }
-}
-
-void
-STDModule::clean()
-{
-  for (int ich = 0; ich < m_size; ++ich) {
-    histvec[ich].clean();
-  }
-}
-
-void
-STDModule::fill(int ch, double value)
-{
-  histvec[ch].fill(value);
-}
-
-void
-STDModule::fill(int ch, int link, double value)
-{
-  histvec[ch + m_index[link]].fill(value);
-}
-
-int
-STDModule::get_local_index(int ch, int link)
-{
-  return ch + m_index[link];
 }
 
 } // namespace dunedaq::dqm
